% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias-estimation.R
\name{estimate_bias}
\alias{estimate_bias}
\title{Estimate bias}
\usage{
estimate_bias(tb, method = "rss", bound = 10)
}
\arguments{
\item{tb}{- data frame with Sample, Taxon, Observed, and Actual}

\item{method}{- "rss" (default) or "center"}

\item{bound}{- lower and upper bound on the log efficiencies for the "rss"
method}
}
\description{
The "center" method should only be used if all samples contain all taxa, and
in that case should give the same results as "rss"
}
