% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ratios}
\alias{plot_ratios}
\title{Inspired by the `boxplot()` method for the `acomp` object of the
`compositions` package.}
\usage{
plot_ratios(.data)
}
\description{
Inspired by the `boxplot()` method for the `acomp` object of the
`compositions` package.
}
