% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tax_labeller}
\alias{tax_labeller}
\title{Formatter for taxon names in the initial form of "Genus_species"}
\usage{
tax_labeller(name)
}
\description{
Formatter for taxon names in the initial form of "Genus_species"
}
