% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias-estimation.R
\name{center}
\alias{center}
\title{Compute the center (compositional mean) of a set of compositions}
\usage{
center(data, weights = rep(1, nrow(data)), method = "proj",
  in_scale = "linear", out_scale = "linear", bound = 10)
}
\arguments{
\item{data}{- data frame or matrix with taxa as columns}

\item{weights}{- sample (row) weights}

\item{method}{- "proj", "gm", or "rss"}

\item{in_scale}{- "linear" (default) or "log"}

\item{out_scale}{- "linear" (default) or "log"}

\item{bound}{- single number giving the lower and upper bound on the alr
efficiencies ("rss" only)}
}
\description{
Compute the center (compositional mean) of a set of compositions
}
